Started working on transitioning cards to mui
