+++ client/src/pages/dashboard/styles/list-card-styles.js
@@ -1,0 +1,25 @@
+import { makeStyles } from "@material-ui/styles";
+
+const useStyles = makeStyles({
+  card: {
+    width: "200px",
+    marginBottom: "10px",
+    marginRight: "20px"
+  },
+  addCard: {
+    display: "flex",
+    width: "200px",
+    height: "258px"
+  },
+  addListCardText: {
+    textAlign: "center"
+  },
+  cardImage: {
+    height: "170px"
+  },
+  content: {
+    textAlign: "center"
+  }
+});
+
+export default useStyles;
