+++ client/src/pages/dashboard/styles/shopping-lists-styles.js
@@ -1,0 +1,61 @@
+import { makeStyles } from "@material-ui/styles";
+
+const useStyles = makeStyles({
+  dashBody: {
+    display: "flex",
+    justifyContent: "center"
+  },
+  newItem: {
+    display: "flex",
+    flexDirection: "column",
+    justifyContent: "center",
+    alignItems: "center",
+    marginTop: "7vh",
+    "& p": {
+      textAlign: "center"
+    }
+  },
+  input: {
+    display: "flex",
+    flexDirection: "row",
+    justifyContent: "space-between",
+    minWidth: "530px",
+    width: "60vw",
+    height: "70px",
+    borderRadius: "50px 50px 50px 50px",
+    backgroundColor: "white",
+    paddingLeft: "30px",
+    "& input": {
+      background: "transparent",
+      width: "25vw",
+      border: "none",
+      fontSize: "20px",
+      color: "rgb(192, 192, 192)",
+      "&:focus": {
+        outline: "none"
+      },
+      "&::placeholder": {
+        color: "rgb(192, 192, 192)"
+      }
+    },
+    "& button": {
+      marginRight: "30px",
+      borderRadius: "50px 50px 50px 50px",
+      width: "170px",
+      margin: "10px 20px 10px 10px"
+    }
+  },
+  verticalLine: {
+    borderLeft: "solid thin rgb(212,212,212)",
+    marginRight: "14px",
+    marginLeft: "-20px"
+  },
+  InputLabel: {
+    width: "100px"
+  },
+  cardsTitleLeft: {
+    display: "flex"
+  }
+});
+
+export default useStyles;
