--- client/src/pages/dashboard/styles/shopping-lists.scss
@@ -1,120 +1,0 @@
-.dash-body {
-  display: flex;
-  justify-content: center;
-
-  &__new-item {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    margin-top: 7vh;
-
-    p {
-      font-size: 24px;
-      font-weight: 500;
-      text-align: center;
-    }
-
-    &__input {
-      display: flex;
-      flex-direction: row;
-      justify-content: space-between;
-      min-width: 530px;
-      width: 60vw;
-      height: 70px;
-      border-radius: 50px 50px 50px 50px;
-      background-color: white;
-      padding-left: 30px;
-
-      input {
-        background: transparent;
-        width: 25vw;
-        border: none;
-        font-size: 20px;
-        color: rgb(192, 192, 192);
-      }
-
-      input:focus {
-        outline: none;
-      }
-
-      ::placeholder {
-        color: rgb(192, 192, 192);
-      }
-
-      .vertical-line {
-        border-left: solid thin rgb(212, 212, 212);
-        margin-right: 14px;
-        margin-left: -20px;
-      }
-
-      .MuiSelect-select {
-        min-width: 69px;
-      }
-
-      .MuiFormControl-root {
-        display: flex;
-        right: 0;
-        flex-direction: row;
-        justify-content: space-between;
-        align-items: space-between;
-      }
-
-      button {
-        margin-right: 30px;
-        border-radius: 50px 50px 50px 50px;
-        width: 170px;
-        margin: 10px 20px 10px 10px;
-      }
-    }
-
-    &__list-cards {
-      display: flex;
-      flex-direction: column;
-      margin-top: 30px;
-      width: 100%;
-      height: 500px;
-
-      .cards-title-left {
-        display: flex;
-      }
-
-      .card {
-        width: 200px;
-        margin-bottom: 10px;
-        margin-right: 20px;
-      }
-
-      .add-card {
-        display: flex;
-        width: 200px;
-        height: 258px;
-
-        .MuiSvgIcon-root {
-          margin-left: 78px;
-          font-size: 50px;
-        }
-
-        .add-list-card-text {
-          text-align: center;
-        }
-      }
-
-      .card-image {
-        height: 170px;
-      }
-
-      .content {
-        text-align: center;
-      }
-
-      // &__card-grid {
-      //   display: grid;
-      //   grid-template-columns: repeat(auto-fill, 224px);
-      //   grid-template-rows: 258px auto 80px;
-      //   grid-column-gap: 10px;
-      //   grid-row-gap: 15px;
-      // }
-    }
-  }
-}
